+++ src/config.ts
@@ -1,0 +1,15 @@
+export const CONFIG = {
+	PROJECT: {
+		NAME: "My Project",
+		VERSION: "1.0.0",
+		DESCRIPTION: "My Project Description",
+		AUTHOR: "My Name",
+		HOMEPAGE: "https://myhomepage.com",
+	},
+	API: {
+		BASE: "/api",
+		ENDPOINTS: {
+			HEALTHCHECK: "/",
+		},
+	},
+};