+++ .github/CHANGELOG.md
@@ -22,6 +22,7 @@ This is the initial version of this template, it does have the following:
 - Added the initial folder structure.
 - Added a healthcheck at `/api` route.
 - Added the common folder structure for the projects, with demo example on each.
+- Added a project configuration file at `src/config.ts`
 
 ### UI and Components
 - We added the light/dark theme to the template.